moved errutils back
